backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
tenacity==8.2.3
requests==2.31.0
python-multipart==0.0.6
jinja2==3.1.2
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
alembic==1.13.0
reportlab==4.0.4
